split LinesTokenizer into tokenizer module
